+++ openmls-wasm/src/lib.rs
@@ -4,11 +4,10 @@ use js_sys::Uint8Array;
 use openmls::{
     credentials::{BasicCredential, CredentialWithKey},
     framing::{MlsMessageBodyIn, MlsMessageIn, MlsMessageOut},
-    group::{config::CryptoConfig, GroupId, MlsGroup, MlsGroupJoinConfig, StagedWelcome},
+    group::{GroupId, MlsGroup, MlsGroupJoinConfig, StagedWelcome},
     key_packages::KeyPackage as OpenMlsKeyPackage,
     prelude::SignatureScheme,
     treesync::RatchetTreeIn,
-    versions::ProtocolVersion,
 };
 use openmls_basic_credential::SignatureKeyPair;
 use openmls_rust_crypto::OpenMlsRustCrypto;
@@ -28,15 +27,6 @@ extern "C" {
 
 /// The ciphersuite used here. Fixed in order to reduce the binary size.
 static CIPHERSUITE: Ciphersuite = Ciphersuite::MLS_128_DHKEMX25519_CHACHA20POLY1305_SHA256_Ed25519;
-
-/// The protocol version. We only support RFC MLS.
-static VERSION: ProtocolVersion = ProtocolVersion::Mls10;
-
-/// The config used in all calls that need a CryptoConfig, using hardcoded settings.
-static CRYPTO_CONFIG: CryptoConfig = CryptoConfig {
-    ciphersuite: CIPHERSUITE,
-    version: VERSION,
-};
 
 #[wasm_bindgen]
 #[derive(Default)]
@@ -93,7 +83,7 @@ impl Identity {
         KeyPackage(
             OpenMlsKeyPackage::builder()
                 .build(
-                    CRYPTO_CONFIG,
+                    CIPHERSUITE,
                     &provider.0,
                     &self.keypair,
                     self.credential_with_key.clone(),
@@ -145,7 +135,7 @@ impl Group {
         let group_id_bytes = group_id.bytes().collect::<Vec<_>>();
 
         let mls_group = MlsGroup::builder()
-            .crypto_config(CRYPTO_CONFIG)
+            .ciphersuite(CIPHERSUITE)
             .with_group_id(GroupId::from_slice(&group_id_bytes))
             .build(
                 &provider.0,