+++ openmls/src/group/tests/test_proposal_validation.rs
@@ -20,7 +20,7 @@ use crate::{
         mls_content::FramedContentBody, validation::ProcessedMessageContent, AuthenticatedContent,
         FramedContent, MlsMessageIn, MlsMessageOut, ProtocolMessage, PublicMessage, Sender,
     },
-    group::{config::CryptoConfig, *},
+    group::*,
     key_packages::{errors::*, *},
     messages::{
         proposals::{AddProposal, Proposal, ProposalOrRef, RemoveProposal, UpdateProposal},
@@ -63,7 +63,7 @@ fn create_group_with_members<KeyStore: OpenMlsKeyStore>(
         provider,
         &alice_credential_with_key_and_signer.signer,
         &MlsGroupCreateConfig::builder()
-            .crypto_config(CryptoConfig::with_default_version(ciphersuite))
+            .ciphersuite(ciphersuite)
             .build(),
         GroupId::from_slice(b"Alice's Friends"),
         alice_credential_with_key_and_signer
@@ -109,7 +109,7 @@ fn new_test_group(
     // Define the MlsGroup configuration
     let mls_group_create_config = MlsGroupCreateConfig::builder()
         .wire_format_policy(wire_format_policy)
-        .crypto_config(CryptoConfig::with_default_version(ciphersuite))
+        .ciphersuite(ciphersuite)
         .build();
 
     (
@@ -357,10 +357,7 @@ fn test_valsem101a(ciphersuite: Ciphersuite, provider: &impl OpenMlsProvider) {
     // a different hpke key, different identity, but the same signature key.
     let dave_key_package = KeyPackage::builder()
         .build(
-            CryptoConfig {
-                ciphersuite,
-                version: ProtocolVersion::default(),
-            },
+            ciphersuite,
             provider,
             &charlie_credential_with_key.signer,
             CredentialWithKey {
@@ -630,7 +627,7 @@ fn test_valsem101b(ciphersuite: Ciphersuite, provider: &impl OpenMlsProvider) {
             provider,
             &alice_credential_with_key.signer,
             &MlsGroupCreateConfig::builder()
-                .crypto_config(CryptoConfig::with_default_version(ciphersuite))
+                .ciphersuite(ciphersuite)
                 .build(),
             GroupId::from_slice(b"Alice's Friends"),
             alice_credential_with_key.credential_with_key.clone(),
@@ -741,10 +738,7 @@ fn test_valsem101b(ciphersuite: Ciphersuite, provider: &impl OpenMlsProvider) {
 
         let dave_key_package = KeyPackage::builder()
             .build(
-                CryptoConfig {
-                    ciphersuite,
-                    version: ProtocolVersion::default(),
-                },
+                ciphersuite
                 provider,
                 &dave_credential_bundle,
             )