+++ openmls/src/treesync/tests_and_kats/tests.rs
@@ -17,9 +17,8 @@ mod test_unmerged_leaves;
 ///   https://github.com/mlswg/mls-protocol/issues/690#issue-1244086547.
 #[apply(ciphersuites)]
 fn that_commit_secret_is_derived_from_end_of_update_path_not_root(ciphersuite: Ciphersuite) {
-    let crypto_config = CryptoConfig::with_default_version(ciphersuite);
     let mls_group_create_config = MlsGroupCreateConfig::builder()
-        .crypto_config(crypto_config)
+        .ciphersuite(ciphersuite)
         .use_ratchet_tree_extension(true)
         .build();
 
@@ -44,7 +43,7 @@ fn that_commit_secret_is_derived_from_end_of_update_path_not_root(ciphersuite: C
         );
         let key_package = KeyPackage::builder()
             .build(
-                CryptoConfig::with_default_version(ciphersuite),
+                ciphersuite,
                 &provider,
                 &credential_with_key_and_signer.signer,
                 credential_with_key_and_signer.credential_with_key.clone(),